moved swarm commands to module
